Making similar changes in HubContextT
