+++ src/Microsoft.AspNetCore.SignalR.Core/HubClients`T.cs
@@ -1,0 +1,52 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microsoft.AspNetCore.SignalR
+{
+    public class HubClients<THub, T> : IHubClients<T>
+    {
+        private readonly HubLifetimeManager<THub> _lifetimeManager;
+
+        public HubClients(HubLifetimeManager<THub> lifetimeManager)
+        {
+            _lifetimeManager = lifetimeManager;
+            All = TypedClientBuilder<T>.Build(new AllClientProxy<THub>(_lifetimeManager));
+        }
+
+        public T All { get; }
+
+        public T AllExcept(IReadOnlyList<string> excludedIds)
+        {
+            return TypedClientBuilder<T>.Build(new AllClientsExceptProxy<THub>(_lifetimeManager, excludedIds));
+        }
+
+        public virtual T Client(string connectionId)
+        {
+            return TypedClientBuilder<T>.Build(new SingleClientProxy<THub>(_lifetimeManager, connectionId));
+        }
+
+        public T MultipleClients(IReadOnlyList<string> connectionIds)
+        {
+            throw new NotImplementedException();
+        }
+
+        public virtual T Group(string groupName)
+        {
+            return TypedClientBuilder<T>.Build(new GroupProxy<THub>(_lifetimeManager, groupName));
+        }
+
+        public T GroupExcept(string groupName, IReadOnlyList<string> excludeIds)
+        {
+            return TypedClientBuilder<T>.Build(new GroupExceptProxy<THub>(_lifetimeManager, groupName, excludeIds));
+        }
+
+        public virtual T User(string userId)
+        {
+            return TypedClientBuilder<T>.Build(new UserProxy<THub>(_lifetimeManager, userId));
+        }
+    }
+}
